requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dateparser==1.1.8
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
paramiko==3.3.1
pycparser==2.21
pycryptodome==3.18.0
PyNaCl==1.5.0
pyTelegramBotAPI==4.13.0
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
websockets==11.0.3
yarl==1.9.2

flask~=2.3.3